requirements.txt
Django==2.2
django-cors-headers==3.7.0
djangorestframework==3.12.4
psycopg2==2.8.6
pytz==2021.1
sqlparse==0.4.1
